Squashed commit of the following:

commit 892e1b978f4e4f73e8fa30d1279ea09b7db2fe32
Author: Michael Altizer <mialtize@cisco.com>
Date:   Mon Jun 29 18:17:40 2020 -0400

    http2_inspect: Make print_flow_issues() regtest-only

commit 63fce83f2b8689c0eb81053c643a5af9123f94a0
Author: Michael Altizer <mialtize@cisco.com>
Date:   Mon Jun 29 18:09:40 2020 -0400

    build: Eradicate u_int usage

    As a bonus, this fixes the Alpine Linux build.

commit 7a1733662671c9e178d7f00e9ce1252df5e8a56a
Author: Michael Altizer <mialtize@cisco.com>
Date:   Mon Jun 29 11:35:41 2020 -0400

    build: Miscellaneous 32-bit build fixes

+++ src/codecs/ip/cd_icmp4.cc
@@ -548,7 +548,7 @@ void Icmp4Codec::log(TextLog* const log, const uint8_t* raw_pkt,
     case icmp::IcmpType::ADDRESSREPLY:
         TextLog_Print(log, "ID: %u  Seq: %u  ADDRESS REPLY: 0x%08X",
             ntohs(icmph->s_icmp_id), ntohs(icmph->s_icmp_seq),
-            (u_int)ntohl(icmph->s_icmp_mask));
+            ntohl(icmph->s_icmp_mask));
         break;
 
     default:
